pyproject.toml
[project]
name = "speech-eureka"
version = "0.1.0"
description = "Modular pipeline for classroom speech transcription, diarization, and speaker identification"
requires-python = ">=3.11"

dependencies = [
    "torch>=2.0",
    "torchaudio>=2.0",
    "transformers>=4.36",
    "pyannote.audio>=3.1",
    "speechbrain>=1.0",
    "hydra-core>=1.3",
    "omegaconf>=2.3",
    "soundfile",
    "gradio>=4.0",
    "jiwer>=3.0",
    "requests",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "ruff",
]

[project.scripts]
speech-eureka = "speech_eureka.main:main"
speech-eureka-gui = "speech_eureka.app:main"
speech-eureka-eval = "speech_eureka.evaluate:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/speech_eureka"]

[tool.uv]
cache-dir = "/workspace/cache/uv"

[[tool.uv.index]]
name = "pytorch-cu124"
url = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = { index = "pytorch-cu124" }
torchaudio = { index = "pytorch-cu124" }

[tool.ruff]
line-length = 100
